Moving the `open` command over to yargs and fixing broken tests.
