+++ cmds/open.js
@@ -1,17 +1,16 @@
 const config = require('config');
 const open = require('opn');
 
-exports.desc = 'Open your current ReadMe project in the browser';
-exports.category = 'utilities';
-exports.weight = 1;
+exports.command = 'open';
+exports.desc = 'Open your current ReadMe project in the browser'.grey;
 
 const configStore = require('../lib/configstore');
 
-exports.run = function({ opts }) {
+exports.handler = function(opts) {
   const project = configStore.get('project');
 
   if (!project) {
-    return Promise.reject(new Error(`Please login using ${config.cli} login`));
+    return Promise.reject(new Error(`Please login using \`${config.cli} login\`.`));
   }
 
   return (opts.mockOpen || open)(config.hub.replace('{project}', project), {
