+++ test/cmds/open.test.js
@@ -1,16 +1,15 @@
 const assert = require('assert');
 const config = require('config');
 
-const configStore = require('../lib/configstore');
-
-const open = require('../cli').bind(null, 'open');
+const configStore = require('../../lib/configstore');
+const open = require('../../cmds/open').handler;
 
 describe('open command', () => {
   it('should error if no project provided', done => {
     configStore.delete('project');
 
     open([], {}).catch(err => {
-      assert.equal(err.message, `Please login using ${config.cli} login`);
+      assert.equal(err.message, `Please login using \`${config.cli} login\`.`);
       return done();
     });
   });
@@ -24,6 +23,6 @@ describe('open command', () => {
       return Promise.resolve();
     }
 
-    open([], { mockOpen });
+    open({ mockOpen });
   });
 });
